ファイル更新

+++ 吉田学園情報ビジネス専門_吉田悠人/作品ファイル/04_二年 後期チーム制作3Dパズル/作品ファイル/プロジェクト/TeamProjects/player.cpp
@@ -282,7 +282,7 @@ void CPlayer::Control(void)
 		// コントローラー十字キーの値の取得
 		CInputJoyStick* pJoyInput = (CInputJoyStick*)CManager::GetInputController();
 		WORD joyCross = CROSS_KEY_NEUTRAL;
-		if (pJoyInput != NULL) joyCross = pJoyInput->GetJoyStick().rgdwPOV[0];
+		if (pJoyInput != NULL) joyCross = (WORD)pJoyInput->GetJoyStick().rgdwPOV[0];
 
 		//エラー音の切替関数
 		ErrorStatus();
